pyproject.toml
[project]
name = "spotify-create-playlist"
version = "0.1.0"
description = "Automatically create a Spotify playlist for a list of tracks"
authors = [
    {name = "beepfish@futurewidgets.eu"}
]
license = {text = "GPLv3"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "spotipy (>=2.25.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "rapidfuzz (>=3.13.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
